requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.2
decorator==4.4.2
dnspython==2.0.0
email-validator==1.1.2

googletrans==3.0.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.8.8
httpcore==0.9.1
httptools==0.1.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
orjson==3.4.6
promise==2.3
pycparser==2.20
pydantic==1.7.3
python-dotenv==0.14.0
python-multipart==0.0.5
python-telegram-bot==12.8
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rfc3986==1.4.0
Rx==1.6.1
six==1.15.0
sniffio==1.1.0
soupsieve==2.0.1
starlette==0.13.6
tornado==6.0.4
ujson==3.2.0
urllib3==1.25.10

watchgod==0.6
websockets==8.1
